+++ src/Endpoint/ALED/Animation/Mask/SlideOff'.hs
@@ -7,7 +7,7 @@ import           Ivory.Language
 import           Ivory.Stdlib
 
 
-renderSlideOff' :: Uint16
+renderSlideOff' :: Uint8
                 -> Sint32
                 -> Record AnimationStruct
                 -> Ivory (AllowBreak (ProcEffects s ())) IFloat
