+++ src/Endpoint/ALED/Animation/Mask/SlideOn'.hs
@@ -7,7 +7,7 @@ import           Ivory.Language
 import           Ivory.Stdlib
 
 
-renderSlideOn' :: Uint16
+renderSlideOn' :: Uint8
                -> Sint32
                -> Record AnimationStruct
                -> Ivory (AllowBreak (ProcEffects s ())) IFloat
