+++ src/Endpoint/ALED/Animation/Mask/SlideOnIn.hs
@@ -7,7 +7,7 @@ import           Ivory.Language
 import           Ivory.Stdlib
 
 
-renderSlideOnIn :: Uint16
+renderSlideOnIn :: Uint8
                 -> Sint32
                 -> Record AnimationStruct
                 -> Ivory (AllowBreak (ProcEffects s ())) IFloat
